internal: move base/support utilities.
