+++ .blueprint/update-generators/generator.ts
@@ -18,7 +18,7 @@
  */
 import { dirname } from 'path';
 import BaseCoreGenerator from '../../generators/base-core/index.js';
-import { createNeedleCallback } from '../../generators/base/support/needles.js';
+import { createNeedleCallback } from '../../generators/base-core/support/needles.js';
 
 export default class UpdateGeneratorsGenerator extends BaseCoreGenerator {
   get [BaseCoreGenerator.WRITING]() {
