+++ generators/base-application/support/index.ts
@@ -23,6 +23,7 @@ export * from './doc.js';
 export * from './enum.js';
 export * from './entity.js';
 export * from './entities.js';
+export * from './faker.js';
 export * from './field-utils.js';
 export { default as prepareEntity } from './prepare-entity.js';
 export * from './prepare-entity.js';
@@ -32,4 +33,5 @@ export * from './prepare-property.js';
 export { default as prepareRelationship } from './prepare-relationship.js';
 export * from './relationship.js';
 export * from './task-type-inference.js';
+export * from '../../base/support/timestamp.ts';
 export * from './update-application-entities-transform.js';
