+++ generators/base-core/generator.ts
@@ -35,7 +35,6 @@ import type Environment from 'yeoman-environment';
 import latestVersion from 'latest-version';
 
 import { CUSTOM_PRIORITIES, PRIORITY_NAMES, PRIORITY_PREFIX, QUEUES } from '../base/priorities.js';
-import { createJHipster7Context, joinCallbacks } from '../base/support/index.js';
 import { CRLF, LF, type Logger, hasCrlr, normalizeLineEndings, removeFieldsWithNullishValues } from '../../lib/utils/index.js';
 
 import type { CascatedEditFileCallback, EditFileCallback, EditFileOptions, ValidationResult, WriteFileOptions } from '../base/api.js';
@@ -60,6 +59,7 @@ import { extractArgumentsFromConfigs } from '../../lib/command/index.js';
 import type GeneratorsByNamespace from '../types.js';
 import type { GeneratorBaseCore } from '../index.js';
 import type { Config as CoreConfig, Features as CoreFeatures, Options as CoreOptions, GenericTaskGroup } from '../base-core/types.js';
+import { createJHipster7Context, joinCallbacks } from './support/index.js';
 import { convertWriteFileSectionsToBlocks } from './internal/index.js';
 
 const {
