+++ generators/base-core/support/index.ts
@@ -16,5 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from './constants.js';
-export * from './timestamp.js';
+export * from './jhipster7-context.js';
+export * from './needles.js';
+export * from './os.js';
+export * from './write-files.js';
