+++ generators/base-core/support/jhipster7-context.ts
