+++ generators/java/support/java-enum.ts
@@ -1,4 +1,4 @@
-import { createNeedleCallback } from '../../base/support/needles.js';
+import { createNeedleCallback } from '../../base-core/support/needles.ts';
 
 type EnumNeedleOptions = { enumName: string; enumValues: string[]; needle?: string; needleValuePrefix?: string };
