+++ generators/spring-cache/generator.ts
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 import BaseApplicationGenerator from '../base-application/index.js';
-import { createNeedleCallback } from '../base/support/needles.js';
+import { createNeedleCallback } from '../base-core/support/needles.ts';
 import writeTask from './files.js';
 import cleanupTask from './cleanup.js';
 import { getCacheProviderMavenDefinition } from './internal/dependencies.js';
