+++ lib/types/application/entity.d.ts
@@ -23,7 +23,7 @@ import type { ServerEntity } from '../../../generators/server/types.js';
 import type { Field as BaseField } from '../base/field.js';
 import type { Relationship as BaseRelationship } from '../base/relationship.js';
 import type { FieldType } from '../../application/field-types.ts';
-import type { FakerWithRandexp } from '../../../generators/base/support/faker.ts';
+import type { FakerWithRandexp } from '../../../generators/base-application/support/faker.ts';
 import type { Field } from './field.js';
 import type { Relationship } from './relationship.js';
